# Label legibility improvements: pitch-scaling and pitched line-following