--- a/src/mbgl/gl/types.hpp
+++ b/src/mbgl/gl/types.hpp
@@ -96,7 +96,7 @@ enum class UniformDataType : uint32_t {
     SamplerCube = 0x8B60,
 };
 
-enum class BufferUsageType : uint32_t {
+enum class BufferUsage : uint32_t {
     StreamDraw = 0x88E0,
     StaticDraw = 0x88E4,
     DynamicDraw = 0x88E8,