… src/mbgl/layout/symbol_projection.cpp
… src/mbgl/programs/symbol_program.hpp
… src/mbgl/programs/uniforms.hpp
… src/mbgl/renderer/buckets/symbol_bucket.cpp
… src/mbgl/renderer/buckets/symbol_bucket.hpp
… src/mbgl/renderer/frame_history.hpp
… src/mbgl/renderer/layers/render_symbol_layer.cpp
… src/mbgl/renderer/layers/render_symbol_layer.hpp
… src/mbgl/renderer/painters/painter_symbol.cpp
… src/mbgl/renderer/render_tile.cpp
… src/mbgl/renderer/render_tile.hpp
… src/mbgl/renderer/tile_pyramid.cpp
… src/mbgl/shaders/collision_box.cpp
… src/mbgl/shaders/symbol_icon.cpp
… src/mbgl/shaders/symbol_sdf.cpp
… src/mbgl/text/collision_feature.cpp
… src/mbgl/text/collision_feature.hpp
… src/mbgl/text/collision_tile.cpp
… src/mbgl/text/collision_tile.hpp
… src/mbgl/text/placement_config.hpp
… src/mbgl/text/quads.cpp
… src/mbgl/text/quads.hpp
… src/mbgl/text/shaping.cpp
… src/mbgl/tile/geometry_tile.cpp
… src/mbgl/tile/geometry_tile.hpp
… src/mbgl/tile/tile.hpp
… src/mbgl/tile/tile_id.hpp
… src/mbgl/util/mat4.cpp
… test/algorithm/update_renderables.test.cpp
… test/programs/symbol_program.test.cpp