Fix #8967 by porting gl-js text pitch scaling behavior to gl-native.

- [x] Calculate shape of collision boxes based on tile distance from camera
- [x] Calculate collision tiles separately for wrapped copies of the same data tile
- [x] Update to new symbol shaders
- [x] Calculate label anchors and pass to updated shaders
- [x] Pass 'max-camera-distance' uniform to updated shaders
- [x] Throttle re-placement during pan operations (currently: change in tile distance always triggers placement, which combined with #8435 can make for label dancing)
- [x] Modify `CollisionFeature` to support greater range of scales implied by pitch scaling
- [x] Selectively bind `fill_color` and `halo_color` vertex attributes to stay under 8 vertex attribute limit
- [x] Update debug collision box code
- [x] Re-enable tests