@ChrisLoer 2017-06-01
@asheemmamoowala Will the de-duplication of wrapped tiles in TilePyramid here play nice with your ImageSource changes?

@jfirebaugh Could you take a look at how I did selective binding `fill_color`/`halo_color`? I don't have it compiling yet on gcc, but I'm very uncertain about the approach I'm using. I think ideally maybe you could have the same vertex buffers but two sets of paint property binders you could pass in to the `draw` call in `painter_symbol.cpp`. I had a lot of trouble trying to tease that logic apart without dramatically changing how the binders work, and that seemed like overkill -- unless "selective binding" is a generalizable idea we'd want to use elsewhere? Still, my current implementation has some code smells: (1) setting a global flag in `Context` to enable the "selective binding", and (2) making the ordering of `IconPaintProperties`/`TextPaintProperties` magical (you need to put the "alternate" property at the end of the list in order to generate the right attribute locations.