@asheemmamoowala 2017-06-02
@ChrisLoer - #8968 doesn't use `OverscaledTileID`, `TilePyramid`, or `update_renderables` so I think your changes to wrapped tiles shouldn't have any impact on ImageSource.