@ChrisLoer 2017-06-02
Thanks @asheemmamoowala!

@ivovandongen This should be ready to test against the ~~mythical~~ 8-attribute device. Just a sanity check that labels have fill and halo colors should be enough, although I've also been testing the second pathway where fill/halo color are defined as a zoom function (had to make a custom style to test that).