@jfirebaugh 2017-06-02
I'd hope to be able to do selective binding without modifying the low level OpenGL wrappers in mbgl/gl/attribute.cpp. The changes here are adding a pretty specialized case to what should be generic code.

I think the core requirement is to enable programs that bind a subset of attributes in a vertex array. [Here's my attempt.](https://github.com/mapbox/mapbox-gl-native/commit/600fc992e62a3d4f262d9fa4d016518d091e16e2) The main blocker is the fact that `Segment` is templated on an `Attributes` type that must be the same as the `Attributes` type used by any `Program` that wants to draw those segments. The reason this is done is so that `Segment` can track the attribute bindings, and if they are different than the previous draw call, update the VAO state. The reason that this must be tracked on a per-`Segment` basis is that the VAO state effectively includes the offset from the beginning of the vertex buffer for that `Segment`.

Options:

* The [ARB_vertex_attrib_binding extension](https://www.khronos.org/opengl/wiki/Vertex_Specification#Separate_attribute_format) allows factoring this offset out. I believe this would allow `Program` to track the binding state instead of `Segment`. This would be ideal, but I doubt the extension is available everywhere.
* Move the tracking of attribute bindings to `Program` anyway, and rebind when the vertex offset changes. This would nullify the performance advantage of VAOs in the case that a vertex buffer has more than 2^16 vertices -- rare, but it happens.
* Allow `Segment` to track binding state for more than one `Program` type. Relax the requirement that its `Attributes` type match the `Program`s `Attributes` -- the `Program`s would merely need to be a subset. I think a previous iteration of VAO binding actually did this -- the moral equivalent of today's `Segment` was templated on the `Program` types that were going to use it, and there was a separate VAO for each one.

The second option sounds best to me right now.