@ivovandongen 2017-06-02
@ChrisLoer The text/halo colors seem to work. But there is some general regressions in rendering on this device that also shows up in label rendering it seems:

![text-colors](https://cloud.githubusercontent.com/assets/64527/26720858/5eed2e54-4792-11e7-8ff8-6c03ebf8b43f.png)
![text-colors-less-ok](https://cloud.githubusercontent.com/assets/64527/26720866/631582b0-4792-11e7-8c52-557804bf4d0a.png)