@ChrisLoer 2017-06-03
@jfirebaugh OK, I took a stab at removing the `Segment` specialization on `Attributes`. I haven't figured out how to re-enable the VAO optimization yet. Is this in the direction you were thinking?