@ansis 2017-06-05
> I think the core requirement is to enable programs that bind a subset of attributes in a vertex array. 

If we implement text and halo rendering in a single draw call is this still necessary? https://github.com/mapbox/mapbox-gl-js/issues/4709#issuecomment-302286933

Are we going to hit the attribute limit for data-driven patterned/dashed lines? Will this require a different solution, like somehow packing multiple style properties in a single attribute?