@ChrisLoer 2017-06-05
> If we implement text and halo rendering in a single draw call is this still necessary?

If we do them in a single draw call, we have to pass both values down, so the selective binding wouldn't be possible any more. We'd have to get around the attribute limit some other way. Looking at your changes in https://github.com/mapbox/mapbox-gl-js/pull/4781, it looks like there are 9 attributes but as much as 72 bits open?

 -  Kind of 16 bits could be put in `a_projected_pos[3]` although the float/int thing is weird as is the dynamic updating.
 - `a_data` has 24 bits free.
 - `a_size` has 32 bits free, I think?