@ansis 2017-06-06
All the pitch-scaling related changes look good to me

> a_size has 32 bits free, I think?

Yep, but since we should move towards treating `a_size` as a regular data-driven property we might not want to pack into it.