@ansis 2017-06-30
@ChrisLoer I've finally ported https://github.com/mapbox/mapbox-gl-js/pull/4781. It took a bit longer than I hoped. Want to start looking over it as I finish the last couple of things?

this fix https://github.com/mapbox/mapbox-gl-native/issues/9272 and https://github.com/mapbox/mapbox-gl-native/issues/9272

remaining things:

- [ ] handle the failing query test
- [ ] fix clang-tidy errors (I'm not seeing any locally, only on CI)

and then there are:
- https://github.com/mapbox/mapbox-gl-js/pull/4853
- https://github.com/mapbox/mapbox-gl-js/pull/4864
- https://github.com/mapbox/mapbox-gl-js/pull/4869

Should these be ported in separate prs?