@ansis 2017-07-06
@ChrisLoer this is ready for merging on my end. See anything else that should be changed?

Should we benchmark on an iOS device before merging? I don't have access to one.