@ChrisLoer 2017-07-07
Benchmarks on my iPhone 5s look fine.

Master:

    Jul  6 19:00:42 Bench GL[1247] <Notice>: Benchmark completed.
    Jul  6 19:00:42 Bench GL[1247] <Notice>: Result:
    Jul  6 19:00:42 Bench GL[1247] <Notice>: Total FPS: 480.4
    Jul  6 19:00:42 Bench GL[1247] <Notice>: Average FPS: 60.0

This branch:

    Jul  6 18:47:45 Bench GL[1218] <Notice>: Benchmark completed.
    Jul  6 18:47:45 Bench GL[1218] <Notice>: Result:
    Jul  6 18:47:45 Bench GL[1218] <Notice>: Total FPS: 480.8
    Jul  6 18:47:45 Bench GL[1218] <Notice>: Average FPS: 60.1