@ChrisLoer 2017-07-07
@ansis I can't approve since I created the PR but your changes look good to me. Assuming you feel good about my changes, let's merge!

One thing that's been bothering me and I've been trying to track down is that the debug collision box colors do not 100% agree with when symbols appear/disappear in pitched views. I thought I had gotten them completely in sync by using the same fade texture, the same label anchor, the same zoom adjustment, etc... I can't find anything you did that would affect it, though.