@ChrisLoer 2017-07-07
Pitch scaling isn't working for icons that are pitched with the map (e.g. one-indicators in streets). I'm having trouble figuring out why, but I'll keep poking into it.