@ChrisLoer 2017-07-07
@ansis I pushed a fix for the icon pitch scaling problem, and bumped the gl-js version to include the enabled tests. We should probably squash the fix commit back in to the others before merging.